+++ http-cache-tower-server/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.2.1] - 2026-01-18
+
+### Changed
+
+- Updated `http-cache` dependency to 1.0.0-alpha.4
+
 ## [0.2.0] - 2026-01-18
 
 ### Changed